Finished 76

+++ app/Http/Controllers/Api/AtendeeController.php
@@ -5,15 +5,27 @@ namespace App\Http\Controllers\Api;
 use App\Http\Controllers\Controller;
 use App\Http\Resources\AttendeeResource;
 use App\Http\Resources\EventResource;
+use App\Http\Traits\CanLoadRelationships;
 use App\Models\Atendee;
 use App\Models\Event;
 use Illuminate\Http\Request;
 
 class AtendeeController extends Controller
 {
+    use CanLoadRelationships;
+
+    private array $relations = ['user'];
+
+    public function __construct()
+    {
+        $this->middleware('auth:sanctum')->except(['index', 'show', 'update']);
+    }
+
     public function index(Event $event)
     {
-        $attendees = $event->attendees()->latest();
+        $attendees = $this->loadRelationships(
+            $event->attendees()->latest()
+        );
 
         return AttendeeResource::collection(
             $attendees->paginate()
@@ -22,24 +34,30 @@ class AtendeeController extends Controller
 
     public function store(Request $request, Event $event)
     {
-        $attendee = $event->attendees()->create([
-            'user_id' => 1,
-        ]);
+        $attendee = $this->loadRelationships(
+            $event->attendees()->create([
+                'user_id' => 1,
+            ])
+        );
 
         return new AttendeeResource($attendee);
     }
 
     public function show(Event $event, Atendee $attendee)
     {
-        return new AttendeeResource($attendee);
+        return new AttendeeResource(
+            $this->loadRelationships($attendee)
+        );
     }
 
 //    public function update(Request $request, string $id)
 //    {
 //    }
 
-    public function destroy(string $event, Atendee $attendee)
+    public function destroy(Event $event, Atendee $attendee)
     {
+        $this->authorize('delete-attendee', [$event, $attendee]);
+
         $attendee->delete();
 
         return response(status: 204);
